small clean-up
